refactor: 카카오 API 호출하는 부분을 다른 클래스로 분리
